+++ Source/Main.java
@@ -165,6 +165,10 @@ public class Main {
                     Console.DEBUG_LOG = true;
 
                     break;
+                case "debug-color":
+                    Console.DEBUG_LOG_COLOR = true;
+
+                    break;
                 case "debug-user-name":
                     if (Argument.check(item, Util.TYPE_STRING))
                         name = item.content;